Cleaned structure
